+++ client/src/components/Carousel/Carousel.module.css
@@ -1,0 +1,43 @@
+.swiper {
+  cursor: grab;
+  background-color: black;
+  width: 85%;
+  margin: 20px auto;
+}
+
+.swiperslide {
+  display: flex;
+  background-color: rgba(255, 255, 255, 0.9);
+  border-radius: 10px;
+  padding: 20px;
+  color: black;
+  font-weight: bold;
+  font-size: 18px;
+  gap: 10px;
+  margin: 20px;
+  max-width: 35rem;
+  min-height: 20rem;
+  max-height: 20rem;
+}
+
+.cards h2 {
+  color: black;
+}
+
+.cards img {
+  max-width: 280px;
+  width: 28vw;
+  height: 300px;
+  object-fit: cover;
+  margin: 0 30px 0 -60px;
+}
+
+.cards button {
+  border: 0;
+  background: var(--primary-color);
+  color: #f8f8f8;
+  padding: 16px 26px;
+  font-size: 16px;
+  border-radius: 40px;
+  width: 14vw;
+}
